수정 : Admin <-> Main Controller 버그 수정
